implementei a página Truck; adicionei página às routes; renomeei a página Trucks para Fleet; corrigi bugs relativos a tipos
