+++ teste-aiko/src/routes.tsx
@@ -1,8 +1,9 @@
 
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Map from "./pages/Map/Map";
-import Trucks from "./pages/Trucks/index";
+import Fleet from "./pages/Fleet/Fleet";
 import MainPage from "./pages/MainPage/MainPage";
+import Truck from "./pages/Truck/Truck";
 
 export default function Router() {
   return (
@@ -11,13 +12,12 @@ export default function Router() {
     <BrowserRouter>
       <Routes>
         <Route path='/' element={<MainPage />}>
-          <Route path='/trucks' element={<Trucks />} />
           <Route index element={<Map/>} />
+          <Route path='/fleet' element={<Fleet />} />
+          <Route path='/truck/:equipmentId' element={<Truck/>} />
         </Route>
       </Routes>
     </BrowserRouter>
     
   )
 }
-
-//<Route path='/categoria/:nomeCategoria' element={<Categoria />} />
